requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.4
regex==2024.5.15
sentence-transformers==2.7.0
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2
nltk==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
rich==13.7.1
orjson==3.10.3
google-generativeai==0.3.2
cryptography==42.0.8
beautifulsoup4==4.12.2
requests==2.32.3
pymongo==4.14.1
motor==3.7.1
beanie==2.0.0
langchain==0.2.12
langchain-community==0.2.12
